Cleaned project
