+++ package.json
@@ -30,6 +30,7 @@
     "homepage": "https://github.com/kaushik-rishi/rocketcp#readme",
     "dependencies": {
         "ejs": "^3.1.6",
+        "fast-diff": "^1.2.0",
         "mkdirp": "^1.0.4",
         "shelljs": "^0.8.4"
     },
